feat: add billing and public reservation foundations
Add the two reviewed database migrations, domain models, EF mappings, billing integration updates, PostgreSQL constraints and concurrency coverage, and the supporting documentation for future feature development.

+++ src/QuanLyChoThuePhongTroWeb.Application/Abstractions/Services/IAiAssistantService.cs
@@ -10,12 +10,12 @@ namespace QuanLyChoThuePhongTroWeb.Application.Abstractions.Services
     {
         Task<List<PhongTroChuaChotRes>> GetPhongTroChuaChotDienNuocAsync(int thang = 0, int nam = 0);
         Task<List<HoaDonChuaThanhToanRes>> GetHoaDonChuaThanhToanAsync(int thang = 0, int nam = 0);
-        Task<double> GetDoanhThuThucThuAsync(DateTime tuNgay, DateTime denNgay);
+        Task<decimal> GetDoanhThuThucThuAsync(DateTime tuNgay, DateTime denNgay);
         
-        Task<List<PhongTrongRes>> GetPhongTrongAsync(double? mucGiaToiDa);
+        Task<List<PhongTrongRes>> GetPhongTrongAsync(decimal? mucGiaToiDa);
         Task<List<HopDongSapHetHanRes>> GetHopDongSapHetHanAsync(int soNgay);
         Task<List<ThongTinKhachThueRes>> GetThongTinKhachThueAsync(string tuKhoa);
-        Task<double> GetCongNoPhongAsync(string soPhong);
+        Task<decimal> GetCongNoPhongAsync(string soPhong);
         Task<List<DoanhThuChiNhanhRes>> GetDoanhThuChiNhanhAsync(int thang = 0, int nam = 0);
         Task<ChiSoDienNuocRes> GetChiSoDienNuocAsync(string soPhong, int thang, int nam);
