+++ src/QuanLyChoThuePhongTroWeb.Application/Abstractions/Services/IVietQRService.cs
@@ -2,7 +2,7 @@ namespace QuanLyChoThuePhongTroWeb.Application.Abstractions.Services
 {
     public interface IVietQRService
     {
-        string GenerateVietQRString(string bankIdOrBin, string accountNumber, double amount, string memo);
+        string GenerateVietQRString(string bankIdOrBin, string accountNumber, decimal amount, string memo);
         byte[] GenerateQRCodePNGBytes(string payload);
     }
 }
