+++ src/QuanLyChoThuePhongTroWeb.Application/Features/HoaDons/Services/IHoaDonCalculatorService.cs
@@ -4,10 +4,10 @@ namespace QuanLyChoThuePhongTroWeb.Application.Features.HoaDons.Services
 {
     public interface IHoaDonCalculatorService
     {
-        (double SoTien, string DienGiai, int SoNgayO) TinhTienPhong(double giaThue, DateTime batDau, DateTime? ketThuc, int thang, int nam);
+        (decimal SoTien, string DienGiai, int SoNgayO) TinhTienPhong(decimal giaThue, DateTime batDau, DateTime? ketThuc, int thang, int nam);
         
-        (double SoLuong, double SoTien, string DienGiai) TinhTienDienNuoc(double chiSoMoi, double chiSoCu, double donGia, string tenDichVu, int soNgayO, int tongNgayTrongThang);
+        (decimal SoLuong, decimal SoTien, string DienGiai) TinhTienDienNuoc(decimal chiSoMoi, decimal chiSoCu, decimal donGia, string tenDichVu, int soNgayO, int tongNgayTrongThang);
         
-        (double SoTien, string DienGiai) TinhTienDichVuCoDinh(double giaDv, int soLuong, string tenDichVu, DateTime batDau, DateTime? ketThuc, int thang, int nam, DateTime? contractStart = null, DateTime? contractEnd = null);
+        (decimal SoTien, string DienGiai) TinhTienDichVuCoDinh(decimal giaDv, decimal soLuong, string tenDichVu, DateTime batDau, DateTime? ketThuc, int thang, int nam, DateTime? contractStart = null, DateTime? contractEnd = null);
     }
 }
