+++ src/QuanLyChoThuePhongTroWeb.Application/Features/HopDongs/DTOs/HopDongKhachThueDetailDto.cs
@@ -14,9 +14,9 @@ namespace QuanLyChoThuePhongTroWeb.Application.Features.HopDongs.DTOs
     public sealed record HopDongDichVuKhachThueDto
     {
         public string tenDichVu { get; init; } = string.Empty;
-        public double donGia { get; init; }
+        public decimal donGia { get; init; }
         public string donVi { get; init; } = string.Empty;
-        public double soLuong { get; init; }
+        public decimal soLuong { get; init; }
     }
 
     public sealed record HopDongKhachThueDetailDto
@@ -24,8 +24,8 @@ namespace QuanLyChoThuePhongTroWeb.Application.Features.HopDongs.DTOs
         public string maHopDong { get; init; } = string.Empty;
         public string thoiDiemBatDau { get; init; } = string.Empty;
         public string thoiDiemKetThuc { get; init; } = string.Empty;
-        public double tienCocPhong { get; init; }
-        public double tienThuePhong { get; init; }
+        public decimal tienCocPhong { get; init; }
+        public decimal tienThuePhong { get; init; }
         public string? tenNguoiDaiDien { get; init; }
         public string? soDienThoaiDaiDien { get; init; }
         public string? cccdDaiDien { get; init; }
