+++ src/QuanLyChoThuePhongTroWeb.Application/Features/LichSuThanhToans/DTOs/LichSuThanhToanDTOs.cs
@@ -11,7 +11,7 @@ namespace QuanLyChoThuePhongTroWeb.Application.Features.LichSuThanhToans.DTOs
         public string TenPhong { get; set; } = string.Empty;
         public string TenNguoiThue { get; set; } = string.Empty;
         public string TenChiNhanh { get; set; } = string.Empty;
-        public double SoTienThanhToan { get; set; }
+        public decimal SoTienThanhToan { get; set; }
         public int PhuongThucThanhToan { get; set; } // 0: Tiền mặt, 1: Chuyển khoản
         public string PhuongThucThanhToanText { get; set; } = string.Empty;
         public string NgayThanhToan { get; set; } = string.Empty;
@@ -21,9 +21,9 @@ namespace QuanLyChoThuePhongTroWeb.Application.Features.LichSuThanhToans.DTOs
 
     public class ThongKeThanhToanRes
     {
-        public double TongDoanhThu { get; set; }
+        public decimal TongDoanhThu { get; set; }
         public int TongSoGiaoDich { get; set; }
-        public double DoanhThuTienMat { get; set; }
-        public double DoanhThuChuyenKhoan { get; set; }
+        public decimal DoanhThuTienMat { get; set; }
+        public decimal DoanhThuChuyenKhoan { get; set; }
     }
 }
