+++ src/QuanLyChoThuePhongTroWeb.Application/Features/LichSuThanhToans/DTOs/LichSuThanhToanKhachThueDto.cs
@@ -8,7 +8,7 @@ namespace QuanLyChoThuePhongTroWeb.Application.Features.LichSuThanhToans.DTOs
         public string? MaGiaoDich { get; init; }
         public string MaHoaDon { get; init; } = string.Empty;
         public string SoPhong { get; init; } = string.Empty;
-        public double SoTienThanhToan { get; init; }
+        public decimal SoTienThanhToan { get; init; }
         public int PhuongThucThanhToan { get; init; }
         public DateTime NgayThanhToan { get; init; }
         public string? GhiChu { get; init; }
