+++ src/QuanLyChoThuePhongTroWeb.Application/Features/PhongTros/DTOs/PhongTroListItemDto.cs
@@ -6,7 +6,7 @@ namespace QuanLyChoThuePhongTroWeb.Application.Features.PhongTros.DTOs
         public string? TenChiNhanh { get; init; }
         public string SoPhong { get; init; } = string.Empty;
         public int TangLau { get; init; }
-        public double GiaThue { get; init; }
+        public decimal GiaThue { get; init; }
         public double DienTich { get; init; }
         public int TrangThai { get; init; }
     }
