+++ src/QuanLyChoThuePhongTroWeb.Application/Features/PhongTros/DTOs/UnpaidInvoiceDto.cs
@@ -6,10 +6,10 @@ namespace QuanLyChoThuePhongTroWeb.Application.Features.PhongTros.DTOs
     public sealed record UnpaidInvoiceDetailDto
     {
         public string tenDichVu { get; init; } = string.Empty;
-        public double donGia { get; init; }
-        public double soLuong { get; init; }
+        public decimal donGia { get; init; }
+        public decimal soLuong { get; init; }
         public string donVi { get; init; } = string.Empty;
-        public double tongTien { get; init; }
+        public decimal tongTien { get; init; }
     }
 
     public sealed record UnpaidInvoiceDto
@@ -18,7 +18,7 @@ namespace QuanLyChoThuePhongTroWeb.Application.Features.PhongTros.DTOs
         public string maHoaDon { get; init; } = string.Empty;
         public int thang { get; init; }
         public int nam { get; init; }
-        public double tongTien { get; init; }
+        public decimal tongTien { get; init; }
         public string tenPhong { get; init; } = string.Empty;
         public string tenNguoiThue { get; init; } = string.Empty;
         public IReadOnlyList<UnpaidInvoiceDetailDto> chiTiets { get; init; } = Array.Empty<UnpaidInvoiceDetailDto>();
